Unified usage of Apache Camel header "ids-header", moved MultiPart Processors to common subproject camel-processors
