+++ camel-processors/src/main/kotlin/de/fhg/aisec/ids/camel/processors/Constants.kt
@@ -20,7 +20,7 @@
 package de.fhg.aisec.ids.camel.processors
 
 object Constants {
-    const val IDSCP2_HEADER = "idscp2-header"
+    const val IDS_HEADER = "ids-header"
     const val IDS_TYPE = "ids-type"
     const val CONTRACT_STORE_KEY = "contract-store-key"
     const val REQUESTED_ARTIFACT_KEY = "requested-artifact"
