+++ camel-processors/src/main/kotlin/de/fhg/aisec/ids/camel/processors/ContractRequestCreationProcessor.kt
@@ -22,7 +22,7 @@ package de.fhg.aisec.ids.camel.processors
 import de.fhg.aisec.ids.api.contracts.ContractConstants.ARTIFACT_URI_PROPERTY
 import de.fhg.aisec.ids.api.contracts.ContractUtils
 import de.fhg.aisec.ids.api.contracts.ContractUtils.SERIALIZER
-import de.fhg.aisec.ids.camel.processors.Constants.IDSCP2_HEADER
+import de.fhg.aisec.ids.camel.processors.Constants.IDS_HEADER
 import de.fraunhofer.iais.eis.Action
 import de.fraunhofer.iais.eis.ContractRequestBuilder
 import de.fraunhofer.iais.eis.ContractRequestMessageBuilder
@@ -45,7 +45,7 @@ class ContractRequestCreationProcessor : Processor {
             if (LOG.isDebugEnabled) {
                 LOG.debug("Serialization header: {}", SERIALIZER.serialize(it.build()))
             }
-            exchange.message.setHeader(IDSCP2_HEADER, it)
+            exchange.message.setHeader(IDS_HEADER, it)
         }
 
         val artifactUri = exchange.getProperty(ARTIFACT_URI_PROPERTY)?.let {
