+++ camel-processors/src/main/kotlin/de/fhg/aisec/ids/camel/processors/ContractRequestProcessor.kt
@@ -22,7 +22,7 @@ package de.fhg.aisec.ids.camel.processors
 import de.fhg.aisec.ids.api.contracts.ContractManager
 import de.fhg.aisec.ids.api.contracts.ContractUtils.SERIALIZER
 import de.fhg.aisec.ids.camel.processors.Constants.CONTRACT_STORE_KEY
-import de.fhg.aisec.ids.camel.processors.Constants.IDSCP2_HEADER
+import de.fhg.aisec.ids.camel.processors.Constants.IDS_HEADER
 import de.fraunhofer.iais.eis.ContractRequest
 import de.fraunhofer.iais.eis.ContractRequestMessage
 import de.fraunhofer.iais.eis.ContractResponseMessageBuilder
@@ -50,7 +50,7 @@ class ContractRequestProcessor(@Autowired private val contractManager: ContractM
         val requestedArtifact = contractRequest.permission[0].target
 
         val contractRequestMessage = exchange.message.getHeader(
-            IDSCP2_HEADER,
+            IDS_HEADER,
             ContractRequestMessage::class.java
         )
 
@@ -60,7 +60,7 @@ class ContractRequestProcessor(@Autowired private val contractManager: ContractM
                 if (LOG.isDebugEnabled) {
                     LOG.debug("Serialization header: {}", SERIALIZER.serialize(it.build()))
                 }
-                exchange.message.setHeader(IDSCP2_HEADER, it)
+                exchange.message.setHeader(IDS_HEADER, it)
             }
 
         val storeKey = exchange.getProperty(CONTRACT_STORE_KEY)?.toString()
