+++ camel-processors/src/main/kotlin/de/fhg/aisec/ids/camel/processors/ResourceUpdateCreationProcessor.kt
@@ -21,7 +21,7 @@ package de.fhg.aisec.ids.camel.processors
 
 import de.fhg.aisec.ids.api.contracts.ContractConstants
 import de.fhg.aisec.ids.api.contracts.ContractUtils.SERIALIZER
-import de.fhg.aisec.ids.camel.processors.Constants.IDSCP2_HEADER
+import de.fhg.aisec.ids.camel.processors.Constants.IDS_HEADER
 import de.fraunhofer.iais.eis.ResourceUpdateMessageBuilder
 import org.apache.camel.Exchange
 import org.apache.camel.Processor
@@ -60,7 +60,7 @@ class ResourceUpdateCreationProcessor : Processor {
                 if (LOG.isDebugEnabled) {
                     LOG.debug("Serialisation header: {}", SERIALIZER.serialize(it.build()))
                 }
-                exchange.message.setHeader(IDSCP2_HEADER, it)
+                exchange.message.setHeader(IDS_HEADER, it)
             }
     }
