+++ examples/example-idscp2/example-idscp2-client-broadcast.xml
@@ -19,9 +19,9 @@
     <camelContext xmlns="http://camel.apache.org/schema/spring">
         <route id="idscp2ReceiverClient">
             <from uri="idscp2client://provider-core:29292?sslContextParameters=#clientSslContext"/>
-            <log message="Client received: ${body} (Header: ${headers[idscp2-header]})"/>
+            <log message="Client received: ${body} (Header: ${headers[ids-header]})"/>
             <!-- Prevents the client consumer from sending the message back to the server -->
-            <removeHeader name="idscp2-header"/>
+            <removeHeader name="ids-header"/>
             <setBody>
                 <simple>${null}</simple>
             </setBody>
