+++ examples/example-idscp2/example-idscp2-client.xml
@@ -44,14 +44,14 @@
             <setBody>
                 <simple>PING</simple>
             </setBody>
-            <setHeader name="idscp2-header">
+            <setHeader name="ids-header">
                 <simple>ping</simple>
             </setHeader>
-            <log message="Client sends: ${body} (Header: ${headers[idscp2-header]})"/>
+            <log message="Client sends: ${body} (Header: ${headers[ids-header]})"/>
             <to uri="idscp2client://consumer-core:29292?awaitResponse=true&amp;idscp2Configuration=#idscp2Config&amp;secureChannelConfigurationBuilder=#tlsConfig"/>
-            <log message="Client received: ${body} (Header: ${headers[idscp2-header]})"/>
+            <log message="Client received: ${body} (Header: ${headers[ids-header]})"/>
             <!-- Prevents the client consumer from sending the message back to the server -->
-            <removeHeader name="idscp2-header"/>
+            <removeHeader name="ids-header"/>
             <setBody>
                 <simple>${null}</simple>
             </setBody>
