+++ examples/example-idscp2/example-idscp2-server-broadcast.xml
@@ -22,10 +22,10 @@
             <setBody>
                 <simple>BROADCAST</simple>
             </setBody>
-            <setHeader name="idscp2-header">
+            <setHeader name="ids-header">
                 <simple>broadcast</simple>
             </setHeader>
-            <log message="Server broadcast to connected clients: ${body} (Header: ${headers[idscp2-header]})"/>
+            <log message="Server broadcast to connected clients: ${body} (Header: ${headers[ids-header]})"/>
             <to uri="idscp2server://0.0.0.0:29292?sslContextParameters=#serverSslContext"/>
         </route>
