+++ examples/route-examples/demo-route.xml
@@ -46,7 +46,7 @@
 					<log message="### Handle ContractAgreementMessage ###"/>
 					<!-- Only processing of incoming request must be performed, no reply -->
 					<process ref="contractAgreementReceiverProcessor"/>
-					<removeHeader name="idscp2-header" />
+					<removeHeader name="ids-header" />
 					<setBody><simple>${null}</simple></setBody>
 				</when>
 				<!-- ArtifactRequestMessage -->
@@ -57,8 +57,8 @@
 				</when>
 				<!-- all other messages -->
 				<otherwise>
-					<log message="### Server received (otherwise branch):\n${body}\n### Header: ###\n${headers[idscp2-header]}"/>
-					<removeHeader name="idscp2-header" />
+					<log message="### Server received (otherwise branch):\n${body}\n### Header: ###\n${headers[ids-header]}"/>
+					<removeHeader name="ids-header" />
 					<setBody><simple>${null}</simple></setBody>
 				</otherwise>
 			</choice>
@@ -147,13 +147,13 @@
 					<to uri="http://172.17.0.2:8080"/>
 					<log message="Response body\n\n${body}"/>
 					<!-- End of message exchange reached -->
-					<removeHeader name="idscp2-header" />
+					<removeHeader name="ids-header" />
 					<setBody><simple>${null}</simple></setBody>
 				</when>
 				<!-- all other messages -->
 				<otherwise>
-					<log message="### Client received (otherwise branch):\n${body}\n### Header: ###\n${headers[idscp2-header]}"/>
-					<removeHeader name="idscp2-header" />
+					<log message="### Client received (otherwise branch):\n${body}\n### Header: ###\n${headers[ids-header]}"/>
+					<removeHeader name="ids-header" />
 					<setBody><simple>${null}</simple></setBody>
 				</otherwise>
 			</choice>
