+++ examples/route-examples/example-idscp2-client.xml
@@ -22,18 +22,18 @@
             <setBody>
                 <simple>PING</simple>
             </setBody>
-            <setHeader name="idscp2-header">
+            <setHeader name="ids-header">
                 <simple>ping</simple>
             </setHeader>
-            <log message="Client sends: ${body} (Header: ${headers[idscp2-header]})"/>
+            <log message="Client sends: ${body} (Header: ${headers[ids-header]})"/>
             <to uri="idscp2client://localhost:29292?connectionShareId=pingPongConnection&amp;sslContextParameters=#clientSslContext"/>
         </route>
 
         <route id="idscp2ReceiverClient">
             <from uri="idscp2client://localhost:29292?connectionShareId=pingPongConnection&amp;sslContextParameters=#clientSslContext"/>
-            <log message="Client received: ${body} (Header: ${headers[idscp2-header]})"/>
+            <log message="Client received: ${body} (Header: ${headers[ids-header]})"/>
             <!-- Prevents the client consumer from sending the message back to the server -->
-            <removeHeader name="idscp2-header"/>
+            <removeHeader name="ids-header"/>
         </route>
 
     </camelContext>
