+++ settings.gradle.kts
@@ -1,7 +1,6 @@
 rootProject.name = "trusted-connector-core"
 
 include(":camel-influxdb")
-include(":camel-multipart-processor")
 include(":camel-processors")
 include(":examples")
 include(":ids-acme")
